Anki add: make Browse sensitive to single/multi card resultset
So, when there are multiple cards, the Anki GUI shows the same set. And
when there's just one card, likewise. This makes it easier to get to the
right card to edit without having to modify the search query in the Anki
GUI.

And clarify the display of "[W]ild: 28 more" in the menu.

And disable display of that once we've invoked it (since there's no 28
more, once you're scrolling through those 28 cards)