Anki add: Remove [F]etch and [R]eview cmds, add cmd [R]eplace
The only reason to use [F]etch before was to see if the online def was
better than my local def and then delete the card and make a new card
with the new content. But, that loses the stats on the age/review
schedule of the card. So, now we just use `update_card` instead. And
there's no need for a separate [F]etch in the UI.